Add new exception and base field classes, remove old model class, and add new model classes and options class
