+++ dj_os/excpetions.py
@@ -1,0 +1,2 @@
+class ConfigurationError(Exception):
+    pass
