+++ dj_os/fields.py
@@ -1,0 +1,2 @@
+class BaseField:
+    pass
