+++ dj_os/models/options.py
@@ -1,0 +1,5 @@
+class ModelOptions:
+
+    def __init__(self, meta) -> None:
+        self.managed = True
+        self.meta = meta
